Move Jewish calendar constants to const file
