+++ tests/components/jewish_calendar/conftest.py
@@ -5,7 +5,7 @@ from unittest.mock import AsyncMock, patch
 
 import pytest
 
-from homeassistant.components.jewish_calendar import config_flow
+from homeassistant.components.jewish_calendar import const
 
 from tests.common import MockConfigEntry
 
@@ -14,8 +14,8 @@ from tests.common import MockConfigEntry
 def mock_config_entry() -> MockConfigEntry:
     """Return the default mocked config entry."""
     return MockConfigEntry(
-        title=config_flow.DEFAULT_NAME,
-        domain=config_flow.DOMAIN,
+        title=const.DEFAULT_NAME,
+        domain=const.DOMAIN,
     )
